Fixing Java style issues in content/
BUG=320711
NOTRY=true

Review URL: https://codereview.chromium.org/111373004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@239989 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/public/test/android/javatests/src/org/chromium/content/browser/test/util/DOMUtils.java
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -140,7 +140,7 @@ public class DOMUtils {
         sb.append("(function() {");
         sb.append("  var node = document.getElementById('" + nodeId + "');");
         sb.append("  if (!node) return null;");
-        sb.append("  if (!node." + fieldName +") return null;");
+        sb.append("  if (!node." + fieldName + ") return null;");
         sb.append("  return [ node." + fieldName + " ];");
         sb.append("})();");
 
@@ -174,9 +174,9 @@ public class DOMUtils {
         Assert.assertNotNull("Failed to get DOM element bounds of '" + nodeName + "'.", bounds);
 
         int clickX = (int) view.getRenderCoordinates().fromLocalCssToPix(bounds.exactCenterX())
-                + (int) view.getContentViewCore().getViewportSizeOffsetWidthPix();
+                + view.getContentViewCore().getViewportSizeOffsetWidthPix();
         int clickY = (int) view.getRenderCoordinates().fromLocalCssToPix(bounds.exactCenterY())
-                + (int) view.getContentViewCore().getViewportSizeOffsetHeightPix();
+                + view.getContentViewCore().getViewportSizeOffsetHeightPix();
         return new int[] { clickX, clickY };
     }
 }
